# api-gateway/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "QuickServe API",
    "description": "API pour l'application QuickServe de transport et services",
    "version": "1.0.0",
    "contact": {
      "name": "Équipe QuickServe",
      "email": "support@quickserve.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Serveur de développement local"
    },
    {
      "url": "https://api.quickserve.com/api",
      "description": "Serveur de production"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Message d'erreur"
          }
        }
      },
...
      "RatingRequest": {
        "type": "object",
        "required": [
          "rating"
        ],
        "properties": {
          "rating": {
            "type": "integer",
            "minimum": 1,
            "maximum": 5,
            "description": "Note (de 1 à 5)"
          },
          "comment": {
            "type": "string",
            "description": "Commentaire optionnel"
          }
        }
      },
...
      "PaymentMethod": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Identifiant unique de la méthode de paiement"
          },
          "type": {
            "type": "string",
            "description": "Type de méthode de paiement (ex. card)"
          },
          "last_digits": {
            "type": "string",
            "description": "Derniers chiffres du numéro de carte"
          },
          "expiry_month": {
            "type": "integer",
            "description": "Mois d'expiration"
          },
          "expiry_year": {
            "type": "integer",
            "description": "Année d'expiration"
          },
          "card_brand": {
            "type": "string",
            "description": "Marque de la carte (ex. Visa, Mastercard)"
          },
          "is_default": {
            "type": "boolean",
            "description": "Indique si c'est la méthode de paiement par défaut"
          },
          "auto_pay": {
            "type": "boolean",
            "description": "Indique si le paiement automatique est activé"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Date d'ajout de la méthode de paiement"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/services/{id}/rate": {
      "post": {
        "tags": [
          "Services"
        ],
        "summary": "Évaluer un service",
        "description": "Donner une note et un commentaire à un service terminé (pour clients)",
        "operationId": "rateService",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identifiant du service",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RatingRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Évaluation enregistrée avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Évaluation enregistrée avec succès"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Données invalides ou service déjà évalué",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Non authentifié",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Accès non autorisé ou service non terminé",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/payments/payment-methods": {
      "post": {
        "tags": [
          "Paiement"
        ],
        "summary": "Ajouter une méthode de paiement",
        "description": "Ajouter une nouvelle méthode de paiement pour le client",
        "operationId": "addPaymentMethod",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "payment_method_id"
                ],
                "properties": {
                  "payment_method_id": {
                    "type": "string",
                    "description": "Identifiant de la méthode de paiement Stripe"
                  },
                  "set_default": {
                    "type": "boolean",
                    "description": "Définir comme méthode par défaut",
                    "default": false
                  },
                  "auto_pay": {
                    "type": "boolean",
                    "description": "Autoriser les paiements automatiques",
                    "default": false
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Méthode de paiement ajoutée avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Méthode de paiement ajoutée avec succès"
                    },
                    "payment_method_id": {
                      "type": "integer",
                      "format": "int64",
                      "description": "Identifiant de la méthode de paiement"
                    },
                    "is_default": {
                      "type": "boolean",
                      "description": "Indique si c'est la méthode par défaut"
                    },
                    "auto_pay": {
                      "type": "boolean",
                      "description": "Indique si le paiement automatique est activé"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Données invalides",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Non authentifié",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Paiement"
        ],
        "summary": "Liste des méthodes de paiement",
        "description": "Récupérer la liste des méthodes de paiement du client",
        "operationId": "getPaymentMethods",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des méthodes de paiement récupérée avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "payment_methods": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/PaymentMethod"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Non authentifié",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/payments/payment-methods/{id}": {
      "delete": {
        "tags": [
          "Paiement"
        ],
        "summary": "Supprimer une méthode de paiement",
        "description": "Supprimer une méthode de paiement existante",
        "operationId": "deletePaymentMethod",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identifiant de la méthode de paiement",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Méthode de paiement supprimée avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Méthode de paiement supprimée avec succès"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Non authentifié",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Méthode de paiement non trouvée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/payments/payment-methods/{id}/set-default": {
      "post": {
        "tags": [
          "Paiement"
        ],
        "summary": "Définir une méthode de paiement par défaut",
        "description": "Définir une méthode de paiement comme méthode par défaut",
        "operationId": "setDefaultPaymentMethod",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identifiant de la méthode de paiement",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Méthode de paiement définie par défaut avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Méthode de paiement définie par défaut avec succès"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Non authentifié",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Méthode de paiement non trouvée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}